requirements.txt
pywebview>=5.1
pythonnet>=3.0.3
pyinstaller>=6.0
scapy
